krackattacks: fix --gtkinit test in 4-way handshake
Commit 9b29501add346d61087869420ae35c35945e412e ("krackattacks: improve
GTK RSC install tests in the 4-way handshake") changed the code of
--gtkinit to perform a completely new 4-way handshake. This was done
because some clients never process a retransmitted Msg3/4. However, the
PTK_RENEW command doesn't support the maxrsc argument, meaning it was
not possible to install the constantly-chaning group key using a very
high replay counter. Instead, the group key was always installed with an
RSC equal to zero. As a result, all devices were considered vulnerable
to the replay of broadcast frames when executing this test.

Fix this behaviour by using retransmitted Msg3/4 instead. The user can
detect whether the client accepts there Msg3/4's by looking at the
output.

Ideally, hostapd is modified to add the maxrsc argument to PTK_RENEW,
but that would require extra testing. Additionally, the PTK_RENEW is
less reliable because a missed frame may cause the connection to fail.

This also fixes #100